[refactor] Add typescript to Notes Context
